Dockerfile
At the moment, this is the skeleton code of what it needs to be. Since the boost library is simple to work with, we have included this here.

Part of a universal SDL3 application is that there are a lot of libraries commonly used, associated with SDL3 that are pretty important. And so I will support these when I can get to them