compile_commands.json
This Intellisense is an autogenerated one that happens during link, and these are just the console commands used to automate the build process.

Update related to a required symlink for linking SDL3 with the general project. Literally what the fuck and why did this have to be so annoying to hit?