Dockerfile
Alright, so this is an insane file that I need to contextualize. So this is for a platform like Linux, which shouldn't have the driver problems??? I couldn't get this container to load in X11 on Mac, so its why my Makefile is a literal container for the ARM Mac OS Architecture. Now I'll check this on Linux eventually, I just am lost in why I'm getting an issue with the renderer part when everything else loads for Linux on Mac?????? Like GLX no context, what the fuck is causing this break????

Now that I have a solution via Makefile, I MIGHT????? be able to get it to translate. run-native works, so, maybe I have a solution present in my mainfile???????