Dockerfile
So, there were a LOTTTTT of libraries required for this thing to even build. Since Alpine's APK manager is DOGSHIT, this means that we need to do a lot of the building ourself. Does this mean a lengthy af build time? Yes, but once we get the libraries linked, we should be able to run the program without much issue.

Curently, we have only fixed the issue of us being able to build via Docker, we cannot run via Docker. However, we can do a native run of the program without problem. And that's because its a failsafe alternative compared to attempting to pipe into X11 via Docker.

Currently, the issue is that there's enough stuff under the hood that still isn't getting properly treated.

For example:
#7 3.982 /usr/glibc-compat/sbin/ldconfig: /usr/lib/libcrypto.so.3 is for unknown machine 183.
#7 3.982
#7 3.982 /usr/glibc-compat/sbin/ldconfig: /usr/lib/libssl.so.3 is for unknown machine 183.
#7 3.982
#7 3.982 /usr/glibc-compat/sbin/ldconfig: /usr/lib/libgomp.so.1.0.0 is for unknown machine 183.
#7 3.982
#7 3.982 /usr/glibc-compat/sbin/ldconfig: /usr/lib/libbrotlicommon.so.1 is for unknown machine 183.
#7 3.982
#7 3.982 /usr/glibc-compat/sbin/ldconfig: /usr/lib/libitm.so is for unknown machine 183.
#7 3.982
#7 3.982 /usr/glibc-compat/sbin/ldconfig: /usr/lib/libsqlite3.so.0.8.6 is for unknown machine 183.

So that means we are pulling these .so files, but it isn't detecting our machine properly.

Nearly everything is getting the git clone treatment, and enough debugging, since often, things happen under the hood that we miss.

Right now, run-native works via Makefile, so I'm less worried about the lack of run-docker. That will be coming, I just finally have a Dockerfile that builds the application without issue. Essentially, all that I should care about for now