Dockerfile
IT FINALLY COMPILES!!!!!

So, you can now run SDL3 with an OpenGL renderer in X11. Huzzah!!!!

So, there's a lot of issues still with things not being substantiated per log file, and I'll tackle these another day. Now, OpenGL works per the container within SDL, and it was hell to try and get this shit to work