Dockerfile - Major Rehaul For Vulkan Render Implementation
Alright, so, we are trying to incorporate Vulkan into SDL3. But we have come across one of the most frustrating things on the planet that requires its own commit as I force Claude to stipulate LLVM15 checks after all packages.

So there is a versioning conflict. Some package was downloaded where It uses the LLVM V15 compiler, while most of my packages are using V16. This wouldn't be such a problem if there were easier ways of quickly checking which package is the problem child.

So we are going to save this, just in case something goes wrong and I need to fetch the previous version