main.hpp
Some of the system headers work, which means that if we were to try to play DOOM in this project, we'd be one step closer to successfully pulling that off.

There's an issue with filio, but apprently this is just ioctl.h with extra steps?????

Lmao headers are very odd inside of containers.