Dockerfile
OFFICIALLY WITH VULKAN WINDOWING SUPPORT!!!!!!!

Since I do not want the same disaster involving LLVM V15 vs V16, this code will be forever written like this. To save myself a game of whackamole, any new libraries we support are going to be very explicitly handled with lots of debugging.

And because of this, we will eventually add some sort of make function to where we solidify the libraries so we can update main without needing to update all the libraries.