Dockerfile
I've added some of the stuff related to SQLite3 that is not working at runtime.

Currently, my focus for today's TODO is to handle anything that's a warning, problem, in my log output. So far, there is a lot of code to cleanup, so I'm taking care of this issue now instead of trying to fix the problem with the header file borking the project. If I can't get past these build issues, then I can't really fix the issue with SQLite. And so hence, my lengthy ass process of diagnosing stuff has begun