CMakeLists.txt
Added much of the needed support involving my libraries built from source.

Currently, there still are a lot of things missing, but I think that has more to do with the Dockerfile over everything. The lack of handling some of the earlier issues compounds into more problems later down the line, at that may be with how I'm attempting to deal with things in Docker.

Since this is a language I am less familiar with, it's up in the air on just how I should fix this, but I'll be starting small and seeing if I can change any of these errors in my Docker log