Dockerfile
Alright, in the process of building out some of the missing dependencies, I got rid of some more LLVM contamination.

And now, we have more and more files missing in the process of building additional things we need. So far, these are more OS specific issues, so understand that the code for this is lengthy so when I write the OS, there is no possibility of some warning flags messing us up