LibrariesDetails.md
Since my Dockerfile can be a big mess, these are some of the critical details surrounding libraries used. This will eventually serve as a good reference point for the Dockerfile, since due to its size, we need a lot of data about paths, or anything that can be helpful in the maintenance of said Dockerfile.

I used this as a means of checking for duplicate libraries, which I removed, and it shaved only a few seconds off of build time.

Since there is so much data associated with my build, this will take quite some time for it to be sufficient enough to guide others.

One of the hardest parts associated with Docker is making sure things are as orgnaized as they can be, as it is really easy to duplicate APKs, or fail to see that a later installed dependency is creating issues with a previous dependency