Dockerfile
I cleaned up a lot of duplicate APKs, make sure it ran, it does. Because this is such a complicated project I want to focus some additional time in improving the starting point to the Alpine build. It has more to do with just seeing if there are any other important dependencies needed that might improve the success rate of SQLite3 from working