Dockerfile
Installed some additional dependencies, shuffled some of the code around so that if there is a warning in the build process, I am not just suddenly wondering if it is just me.

So far, this project is going extremely well!!!!!!!

I still have 106 'not found' error messages, but now I can finally tackle these problems, or, I should be able to tackle these things better now.

It felt like a wasted day, but the code looks and is so much better