create-filesystem.sh
Because I am trying to have a filsystem closer to an OS, this means that we will need to slowly port over where our code is being stored. Now, I want to do this because what ends up becoming a problem is that since we are shooting in the dark on where things end up, it means we write a lot of pointless code about the paths of my files.

Now, this is going to be a complex endeavor, since so much of my code is based around assumed paths. Part of what will make this project work better is that I will have a place to store shell scripts, which means that I can write a lot more debug functionality depending on need and have it stored conveniently