create-filesystem.sh
So, my plans for this is to verify EVERY little file installed. The reason for this has to do with more documentation reasons. It might be useful to have a library list, but it is more useful to have a file list, something that will document how the filesystem works and settle any potential problems.

I usually wouldn't want to do a big rewrite, but given I've been unable to solve a lot of foundational problems, it's time to refactor and settle the biggest problems.

This is where build times are about to go fucking insane, because I am doing double the amount of building