Dockerfile
We now have a system to check for any missing files, via shell script. This is one of the biggest debug tools I've ever used so far. So now, I can go to bed knowing I'll have a series of new stuff to look at tomorrow.

Finally, we can fix a lot of the file dependency problems that have been problematic for the past few days