check-filesystem.sh
We use this at the end of each section. We basically want to see ALL the files that have been generated, so I can create an in depth filestructure reference. Because its easy to get lost on what paths are used, this is meant as a shell code to peer into my OS.