env_tracer.sh
Alright, one of the frustrating parts about having so many environmental variables is that shit can go fucky-wucky. And so if all of a sudden, we download a package where previous environmental variables are shifted to a new place, we need to see just exactly where this happens. These environment changes can often be a problem, and hence, this shell code