stage_went_wrong_tracker.sh
So when there's a shift from stage to stage, it is often hard to tell if things finalized correctly. So if there's discrepancies between each stage, we need to know what is happening since you really don't get that data from under the hood.

So being able to track things when a stage changes, is integral of making sure we didn't miss anything that could cause further issue