dep_chain_visualizer.sh
So you know how there's a big issue involving knowing what belongs inside of each dependency. Well often it is tough to get this to work inside of Docker without it being a MILLION LINES OF CODE. And so to save our sanity of trying to grep all and everything, it is much better to have this done via shell code instead.

All these various issues is why I am focused on improving my debug tools, because attempting to do this blind just has so many problems.