layer-bloat_analyzer.sh
While we do check for compiler bloat, we need the ability to also check for library bloat. Chances are, duplicate libraries won't create the problems, it just is great to make sure I'm not shoving too many packages in the OS.

Embedded systems need to utilize memory most efficiently, and duplicate packages just take up precious memory. Because if I can make this OS 2GB, that would be great!