sgid_suid_scanner.sh
This might be rare because I don't really tackle any permissions problems, but this is better for when we work with the kernel. Esp if we are trying to OS development something and need basic security methods.

I haven't yet got into security yet, but this just exists for the sake of a later date