src_fetch.sh
Welcome to the autosuggest of Docker, which does exist, and does provide you detail on exactly alternatives.

This is so you aren't needing to skim around the thousands of Alpine packages. Have you ever tried figure out alternatives? Yeah, it's a fucking bitch to do it, not easy in the slightest.