Dockerfile - DO NOT DOWNLOAD YET (download OldDockerfile.md instead)
Alright, if somebody has managed to find this repo and download it, you will find that there are so many issues compiling.

Yes, that is the intent. At the moment, I do not care if there are this many issues for one reason. This is a complete rewrite.

Yes, this entire Dockerfile is rewritten. And so because it is so new, so many issues. So why rewrite the Dockerfile? Well due to the amount of hidden bugs I was getting because things weren't set up well, I needed a better system to try even if it costs me a lot more development time.

To save my sanity from the constant issues of dependencies having issues, and not being able to see under the hood, this is why the project has gotten much more shell heavy. Because as soon as I have an issue that can't be simply resolved, there needs to be better means of fixing the problem in full without missing something.

Since this project is on the relatively small side, and I've been trying to make a decent back end, it was much better to start fixing the problem now than when we have even MORE packages later down the line.

Apparently embedded is a lot more storage friendly???? So, with that, I need to be making use of it well before I can even worry about OS size.

So now, our filesystem and contamination commands are much better at finding problems. This will be the same hunt of trying to enforce LLVM16, but now that I've done it so much, it'll be much easier to get rid of this problem. And if I can't fix that, then I'll shellcode the problem to death