Dockerfile - Checksum Problem
Alright, so I have started to narrow down the issues, and its just a lack of certain dependencies that is causing issue. We are moving in the correct direction because my logging file gets longer and longer, and that's because we reward successes. There's a checksum issue, but LESS of a checksum issue