Dockerfile
I HAVE FOUND THE AREA IN WHICH IT FUCKING CRASHES WHICH IS GREAT BECAUSE I LITERALLY COULDN'T FIGURE OUT SHITTTTTTTTTT.

Do you know how much time I spent trying to find where the bug was occuring????

Literally when I copy the OS, that is where the binary dissapears. Its probably something to do with a wrong path in my CMake file????

Jfc this filesystem issue has a skill floor increase of like 50 points. I'll take the suffering now if it means solving the problem in a few extra longer days