Dockerfile
Filesystem layout at the beginning so we don't feel lost in our Dockerfile as it scales.

Think of it as a lengthy ass directory on everything and where it needs to go. Added back some of the core essentials back, which are important to the graphics stack.

I am planning on approaching this much differently. Since it feels like I have some folder bloat, I'm going to be much better about where I place APKs, and that organization will be useful so we don't do something like make a binary unlivable.

WE ARE SO FUCKING BACK