Dockerfile
Added some more essential libraries back. Libraries are moved to their proper folders, much better compared to stuffing these folders with a million different librares.

I am slowly going to integrate things back, so we don't get the problem of last time of fucking breaking my code. It is going to take about 2 weeks to get my code back to where it once was, but it was well worth doing this