Dockerfile
So there's so many issues under the hood. MESA fails, this code wouldn't even properly create a window if I tried. But the point is, that I do not get a lack of building the app and a critical failure. It took about 3 days to move all the build from source libraries over with how many there were.

Now, the task is to make sure that I don't have any non-fatal errors, since I feel that is most important since my SDL3 code is ready to go