CMakeLists.txt
libepoxy incorporated, still about 18 build from source libraries that need to be stringently checked for whether or not it fails the CMake file. I am not taking anymore chances this time with assuming the code is good when I attempt to replicate it via DeepSeek (because I fucking cannot stand writing boilerplate, and as long as we aren't fucking over the project, I will accept it, but that's only because I am trying to set things up for the sake of viewing logs).

This code is the type of annoying this to replicate and I just wanna tackle more important issues like why my Dockerfile has so many issues under the hood.