CMakeLists.txt
Yeah so this works and I am going to bed. Yes, we absolutely need a fallback method if we are going to handle CMake's strict standard and Dockerfile's strict standard. I'm not playing this referee game in which I try to localize things so that my program doesn't have a bitch fit.

Now, we can dump our CMake logs and also compile. Which means all we can focus on in the morning is fixing some big problems. I've triple marked this code because this is something that can't be touched. Well, I mean, we are going to expand upon it at some point, but for now, all I care for is just making sure the backend works how it should before even EXANDING when it comes to files.

There's a lot of really important parts of the project, this is one of the most important recent developments. It's why I've marked it the way that I did, because this is the core in how I need the codebase to funciton, and it cannot be modified to disclude the other