main.cpp
HORAY!!!!! SDL3 WORKS IN ITS MOST PRIMITIVE WAY

So, just so you understand why this is so important, this means that we were able to create an SDL3 context. This might be small, a window context, but that usually means I can at least create a basic window if that happens.

There are only 2 more tests I am doing tonight, because if all goes well, we should have an XQuartz context. I'm still far away from restoring things in full, but I'll take the minimum if it means a window