Manifest_Standards.md
These are the standards we must adhere to involving manifest files. Albeit, with tweaks since CMake still is struggling to realize WHERE my paths for something as basic as FOP is. Idk why there is just such fucking issues involving this, but, here we are.

Also, consider it the method that we also make sure anything that struggles with being built in the debug stage gets addressed at this point. SO this should at least be useful for Docker if we still aren't getting shit with CMake