Dockerfile
Still no luck with trying to get things to work as they should. We have issues trying to get things to work for the debug phase, but so far, in the build phase, things are going fine.

These are the types of issues you deal with that take a long time to solve exactly for this reason. I'll diagnose this after class