TODO.md
Right now, I will be ONLY focusing on libraries within my build stage. AFTER, we can deal with issues in the debugger stage, since this is gonna require a lot of diagnosis. For now, we are close to resolving all the build problems, and so, in about two weeks, I should be ready to fix these issues