TODO.md
We are so back. Currently, we have fixed a major issue, libdrm, with an amazing standard I'll be using on the rest of the broken stuff.

Because of this newest fix, this means we hold part of the solution to fixing everything that has been really difficult to fix without breaking the entire code.

Even if I solve all the current TODO problems, there still will be a lot of other fixes needed. My goal is to get everything done that'll allow for OpenGL and Vulkan rendering to work. So far, there are still too many broken elements for this to happen, which are the few things I'm patching up