apk-retry.sh
Designed to deal with hangups via creating a modification to apk add, where we wrap it with apk-retry instead of standard apk. Since our build from source is unlikely to deal with hangups as frequently, we do this to fix the issue of packages freezing and do a series of retries. We give 5 just in case