Dockerfile
Shaderc currently needs glslang, so we will most likely need to build from source so we can chuck the libraries and binaries into shaderc. Unless there is some other method that I am not aware of that can have us dump glslang from standard apk add. It's like a fight with every single library, but if this is what I need to do to resolve my code issues, I'll do it

SPIRV-Tools gets moved into the third-party folder correctly, and it is gross. I absolutely despise the workaround, but, if it means getting shaderc to FINALLY link, I'll take it. We basically have to force third-party to accept our library and binaries, and so any dependency will need to be passed through this way.

We are near complete with shaderc, and I'll resolve the problem tonight by doing a build from source construction. Currently, none of glslang is found like it should, so either I've messed up with sysroot population, or I need to build from source. I just hate building from source anything, I DESPISE IT