Makefile
After a long break and coming back to this project, I've decided to work on just only bringing the minimum of what is needed instead. And plus now, we finally have an Alpine Linux version of SDL3 so I don't have to rip my hair out anymore.

So now, because I do not need a million complicated packages to download, multi-stage builds because I don't need to build from source, this file becomes VERY simple