APKBUILD
Alright, usually, you'd need more packages, but since it doesn't like having two locations worth of packages, you have to widdle it down. Frustrating, because then I can't check true dependencies. So given that this is a package for my own project, it'll do, but for other people's projects, I'll need to do this test in a barebones Dockerfile