fb-wrapper.sh
So here is an interesting thing about the frame bufferer, it is a pick little shit, and because we do a working fbConfig and glx, it does not satisfy the evil. SO we have to get rid of these error logs.

HOWEVER, it probably is the reason why I can't bind to certain paths, so the question is how i can manipulate bash so I find all the addresses to the drivers we need to access.