Dockerfile
I do not get how in literally the middle of my session, a library all of a sudden wanted to pose a conflict?

I'm not a fan of force-overwrite, it just feels like cheating. ALSO , aren't there like so many libraries that use the same thing so its like, okay but why does sqlite need to be the problem child here?

Also, I do not get how the mp3 codec is failing when everything is doing just fine.